Moving tests out to separate module
